Change tonnes int to double

+++ EPRN.Common.Data/DataModels/PackagingRecoveryNote.cs
@@ -18,7 +18,7 @@ namespace EPRN.Common.Data.DataModels
 
         public string SentTo { get; set; }
 
-        public int? Tonnes { get; set; }
+        public double? Tonnes { get; set; }
 
         public int? SiteId { get; set; }
     }
